added ediror per project instance
